rename config_strs -> cfg_strs
