+++ slk/config/cfg_strs.py
